Migrate handler for GenericOptionsHandler
